+++ test_scripts/Polices/App_Permissions/176_P_ATF_Steal_focus_validation_false_PTU.lua
@@ -1,0 +1,139 @@
+-- Requirement summary:
+-- [[Policies]] <app_id> policies and "steal_focus" validation.
+
+-- Description:
+-- In case the <app id> policies are assigned to the application, PoliciesManager must validate "steal_focus" section and in case "steal_focus:true",
+-- PoliciesManager must allow SDL to pass the RPC that contains the soft button with STEAL_FOCUS SystemAction.
+-- Note: in sdl_preloaded_pt. json, should be "steal_focus:false" for Policies.
+-- Note: in ptu.json, should be "steal_focus:false".
+
+-- 1. RunSDL. InitHMI. InitHMI_onReady. ConnectMobile. StartSession.
+-- 2. Activiate Application for allow sendRPC Alert
+-- 3. Run Policy Update with steal_focus true value for Current_App
+-- 4. MOB-SDL: SendRPC with soft button, STEAL_FOCUS in SystemAction
+-- Expected result
+-- SDL must response: success = false, resultCode = "DISALLOWED"
+-------------------------------------------------------------------------------------------------------------------------------------------------------
+--[[ General configuration parameters ]]
+config.deviceMAC = "12ca17b49af2289436f303e0166030a21e525d266e209267433801a8fd4071a0"
+--ToDo: shall be removed when issue: "ATF does not stop HB timers by closing session and connection" is fixed
+config.defaultProtocolVersion = 2
+
+--[[ Required Shared libraries ]]
+local commonFunctions = require ('user_modules/shared_testcases/commonFunctions')
+local commonSteps = require ('user_modules/shared_testcases/commonSteps')
+
+--[[ General Precondition before ATF start ]]
+commonSteps:DeleteLogsFileAndPolicyTable()
+
+--[[ General Settings for configuration ]]
+Test = require('connecttest')
+require('cardinalities')
+require('user_modules/AppTypes')
+
+--[[ Preconditions ]]
+function Test:Precondition_ActivateApplication()
+
+  local RequestId = self.hmiConnection:SendRequest("SDL.ActivateApp", {appID = self.applications["Test Application"]})
+  EXPECT_HMIRESPONSE(RequestId, { result = {code = 0, isSDLAllowed = false}, method = "SDL.ActivateApp"})
+  :Do(function(_,_)
+      local RequestId1 = self.hmiConnection:SendRequest("SDL.GetUserFriendlyMessage", {language = "EN-US", messageCodes = {"DataConsent"}})
+      EXPECT_HMIRESPONSE(RequestId1,{result = {code = 0, method = "SDL.GetUserFriendlyMessage"}})
+      :Do(function(_,_)
+          self.hmiConnection:SendNotification("SDL.OnAllowSDLFunctionality", {allowed = true, source = "GUI", device = {id = config.deviceMAC, name = "127.0.0.1"}})
+          EXPECT_HMICALL("BasicCommunication.ActivateApp")
+          :Do(function(_,data)
+              self.hmiConnection:SendResponse(data.id,"BasicCommunication.ActivateApp", "SUCCESS", {})
+              EXPECT_NOTIFICATION("OnHMIStatus", {hmiLevel = "FULL", systemContext = "MAIN"})
+            end)
+          EXPECT_NOTIFICATION("OnPermissionsChange", {})
+        end)
+    end)
+end
+
+function Test:Preconditions_Update_Policy_With_Steal_Focus_FalseValue_for_Current_App()
+  local RequestIdGetURLS = self.hmiConnection:SendRequest("SDL.GetURLS", { service = 7 })
+  EXPECT_HMIRESPONSE(RequestIdGetURLS,{result = {code = 0, method = "SDL.GetURLS", urls = {{url = "http://policies.telematics.ford.com/api/policies"}}}})
+  :Do(function()
+      self.hmiConnection:SendNotification("BasicCommunication.OnSystemRequest",
+        {
+          requestType = "PROPRIETARY",
+          fileName = "filename"
+        }
+      )
+      EXPECT_NOTIFICATION("OnSystemRequest", { requestType = "PROPRIETARY" })
+      :Do(function()
+          local CorIdSystemRequest = self.mobileSession:SendRPC("SystemRequest",
+            {
+              fileName = "PolicyTableUpdate",
+              requestType = "PROPRIETARY"
+            }, "files/ptu_general_steal_focus_false.json")
+          local systemRequestId
+          EXPECT_HMICALL("BasicCommunication.SystemRequest")
+          :Do(function(_,data)
+              systemRequestId = data.id
+              self.hmiConnection:SendNotification("SDL.OnReceivedPolicyUpdate",
+                {
+                  policyfile = "/tmp/fs/mp/images/ivsu_cache/PolicyTableUpdate"
+                })
+              local function to_run()
+                self.hmiConnection:SendResponse(systemRequestId,"BasicCommunication.SystemRequest", "SUCCESS", {})
+              end
+              RUN_AFTER(to_run, 800)
+              self.mobileSession:ExpectResponse(CorIdSystemRequest, {success = true, resultCode = "SUCCESS"})
+              EXPECT_HMINOTIFICATION("SDL.OnStatusUpdate", {status = "UP_TO_DATE"})
+            end)
+        end)
+    end)
+end
+
+--[[Test]]
+commonFunctions:newTestCasesGroup("Test")
+function Test:TestStep_SendRPC_with_STEAL_FOCUS_Value()
+  local CorIdAlert = self.mobileSession:SendRPC("Alert",
+    {
+      alertText1 = "alertText1",
+      alertText2 = "alertText2",
+      alertText3 = "alertText3",
+      ttsChunks =
+      {
+        {
+          text = "TTSChunk",
+          type = "TEXT",
+        }
+      },
+      duration = 5000,
+      playTone = true,
+      progressIndicator = true,
+      softButtons =
+      {
+        {
+          type = "TEXT",
+          text = "Keep",
+          isHighlighted = true,
+          softButtonID = 4,
+          systemAction = "STEAL_FOCUS",
+        },
+
+        {
+          type = "IMAGE",
+          image =
+          {
+            value = "icon.png",
+            imageType = "DYNAMIC",
+          },
+          softButtonID = 5,
+          systemAction = "STEAL_FOCUS",
+        },
+      }
+    })
+  EXPECT_RESPONSE(CorIdAlert, { success = false, resultCode = "DISALLOWED"})
+end
+
+--[[ Postconditions ]]
+commonFunctions:newTestCasesGroup("Postconditions")
+function Test.Postcondition_StopSDL()
+  StopSDL()
+end
+
+return Test
